+++ docs/_includes/eleventyComponents.js
@@ -14,21 +14,20 @@
  *   limitations under the License.
  */
 
-const { Compile } = require("../../");
+const { Source } = require("../../");
 const site = require("../_data/site");
 
-module.exports.Log = (env, props, _children) => {
+const Log = Source.func("Log", (env, props, _children) => {
   console.log(props);
   return env.return("");
-};
+});
 
-module.exports.Debugger = (env, _props, _children) => {
+const Debugger = Source.func("Debugger", (env, _props, _children) => {
   debugger;
   return env.return("");
-};
+});
 
-const footerAst = Compile.makeAst(
-  `<footer class="footer">
+const footerSrc = `<footer class="footer">
   <p>
     Published in {{ year }} by
     {%~ match link with null ~%} {{ name }}
@@ -40,26 +39,31 @@ const footerAst = Compile.makeAst(
     <a href="{{ siteUrl }}/license/">View the license</a>.
   </p>
 </footer>
-`,
-  "Footer"
-);
+`;
 
-module.exports.Footer = (env, props, children) => {
-  if (!props.year) {
-    props.year = new Date().getFullYear();
+const Footer = Source.funcWithString(
+  "Footer",
+  footerSrc,
+  (ast) => (env, props, children) => {
+    if (!props.year) {
+      props.year = new Date().getFullYear();
+    }
+    return env.render(ast, props, children);
   }
-  return env.render(footerAst, props, children);
-};
-
-const linkAst = Compile.makeAst(
-  `<a href="{{ href }}" aria-current="{{ current }}">
-  {{ Children }}
-</a>`,
-  "Link"
 );
 
-module.exports.Link = (env, { path, page }, children) => {
-  const current = path === page.url ? "true" : "false";
-  const href = site.url + path;
-  return env.render(linkAst, { href, current }, children);
-};
+const linkSrc = `<a href="{{ href }}" aria-current="{{ current }}">
+  {{ Children }}
+</a>`;
+
+const Link = Source.funcWithString(
+  "Link",
+  linkSrc,
+  (ast) => (env, { path, page }, children) => {
+    const current = path === page.url ? "true" : "false";
+    const href = site.url + path;
+    return env.render(ast, { href, current }, children);
+  }
+);
+
+module.exports = [Log, Debugger, Footer, Link];